Writer assign incrementing id to events before append
